postgreStructure/requirements.txt
absl-py==0.7.1
alembic==1.0.8
astor==0.7.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
flask-marshmallow==0.10.0
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gast==0.2.2
grpcio==1.19.0
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Mako==1.0.8
Markdown==3.1
MarkupSafe==1.1.1
marshmallow==2.19.1
marshmallow-sqlalchemy==0.16.1
mock==2.0.0
numpy==1.16.2
pbr==5.1.3
Pillow==5.4.1
prettytable==0.7.2
protobuf==3.7.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1
requests==2.21.0
scipy==1.2.1
six==1.12.0
SQLAlchemy==1.3.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.15.1
